Update rgi20.md (#35)
I left it here but tried to avoid the word ´add a region´ since technical are was not added by this move since 19-31 existed, but left it here but tried to explain and also give a motivation for the change. going  from 19 to 20 regions is a major move that folks notice so folks will wonder why, esp if it makes no difference technically at the moment with nothing mapped.